Squished minimal.yaml comments and object references
