+++ configurations/minimal.yaml
@@ -1,97 +1,72 @@
-# File metadata, to be printed by the simulator
-name:
+name: # File metadata, to be printed by the simulator
   "Minimal Dispatch Example"
 
-# Defines the bounds for case and ambulance random location initialization
-location_generator:
+hospitals: # Defines 2 hospitals in the UCSD area
+  class: LocationSet
+  classpath: ems.datasets.location.location_set
+  latitudes: [32.879802, 32.888216]
+  longitudes: [-117.236564, -117.239697]
+
+bases: # Defines 6 bases in the UCSD area
+  class: LocationSet
+  classpath: ems.datasets.location.location_set
+  latitudes: [32.877328, 32.873803, 32.876727, 32.875811, 32.884714, 32.888589]
+  longitudes: [-117.234695, -117.239527, -117.242650, -117.230869, -117.236528, -117.242113]
+
+ambulances: # Defines 6 ambulances stationed at each of the 6 bases
+  class: CustomAmbulanceSet
+  classpath: ems.datasets.ambulance.custom_ambulance_set
+  bases: "@bases"
+
+location_generator: # Generates case locations uniformly in a circle of 2 kilometers centered on UCSD
   class: CircleLocationGenerator
   classpath: ems.generators.location.circle
-
-  # Defines coordinates for the center of the circle
   center_latitude: 32.880506
   center_longitude: -117.231017
-
-  # Defines the radius of the circle (in km)
   radius_km: 2
 
-# Defines the duration spent travelling, at incident, and at hospital
-duration_generator:
+duration_generator: # Generates durations spend travelling, at incident, and at hospital uniformly within 10 to 20 minutes
   class: RandomDurationGenerator
   classpath: ems.generators.duration.random_duration
-
-  # Bounds in minutes for durations
   lower_bound: 10
   upper_bound: 20
 
-# Predefines hospitals
-hospitals:
-  class: LocationSet
-  classpath: ems.datasets.location.location_set
+hospital_selector: # Algorithm selects a random hospital to deliver a patient after an incident
+  class: RandomHospitalSelector
+  classpath: ems.algorithms.hospital_selectors.select_random
+  hospital_set: "@hospitals"
 
-  # Defines latitudes
-  latitudes: [32.879802, 32.888216]
+event_generator: # Generates random events for each case
+  class: EventGenerator
+  classpath: ems.generators.event.event_generator
+  travel_duration_generator: "@duration_generator"
+  incident_duration_generator: "@duration_generator"
+  hospital_duration_generator: "@duration_generator"
+  hospital_selector: "@hospital_selector"
 
-  # Defines longitudes
-  longitudes: [-117.236564, -117.239697]
+case_time_generator: # Generates case times; interval between cases are uniformly generated between 8 to 10 minutes
+  class: RandomCaseTimeGenerator
+  classpath: ems.generators.time.random_time
+  lower_bound: 8
+  upper_bound: 18
 
-# Defines the simulator to be used
-simulator:
+cases: # Defines 500 cases starting on January 31, 2019; randomly generated durations, locations, and times
+  class: RandomCaseSet
+  classpath: ems.datasets.case.random_case_set
+  initial_time: 2019-1-31T00:00:00.1Z
+  quantity: 500
+  case_time_generator: "@case_time_generator"
+  case_location_generator: "@location_generator"
+  event_generator: "@event_generator"
+
+ambulance_selector: # Algorithm to randomly select an ambulance upon dispatch to a case
+  class: RandomSelector
+  classpath: ems.algorithms.selection.dispatch_random
+
+simulator: # Simulation to run
   class: EventDispatcherSimulator
   classpath: ems.simulators.event_simulator
-
-  # Allows for progress output
   debug: True
-
-  # Defines the ambulances to be used in the simulation
-  ambulances:
-    class: RandomAmbulanceSet
-    classpath: ems.datasets.ambulance.random_ambulance_set
-
-    # Number of ambulances to generator
-    count: 6
-    base_generator: "@location_generator"
-
-  # Defines the cases for the simulation
-  cases:
-    class: RandomCaseSet
-    classpath: ems.datasets.case.random_case_set
-
-    # Starting time
-    initial_time: 2019-1-31T00:00:00.1Z
-
-    # Number of cases to generate
-    quantity: 50
-
-    # Defines method on when cases are randomly generated
-    case_time_generator:
-      class: RandomCaseTimeGenerator
-      classpath: ems.generators.time.random_time
-
-      # Duration in minutes between each case
-      lower_bound: 8
-      upper_bound: 18
-
-    # Bounds for case locations
-    case_location_generator: "@location_generator"
-
-    # Defines how the events in the case should be generated (e.g. TO HOSPITAL, TO INCIDENT, etc.)
-    event_generator:
-      class: EventGenerator
-      classpath: ems.generators.event.event_generator
-
-      # Duration generators
-      travel_duration_generator: "@duration_generator"
-      incident_duration_generator: "@duration_generator"
-      hospital_duration_generator: "@duration_generator"
-
-      # Hospital Selection
-      hospital_selector:
-        class: RandomHospitalSelector
-        classpath: ems.algorithms.hospital_selectors.select_random
-
-        hospital_set: "@hospitals"
-
-  # Defines algorithm for selecting ambulances during dispatch
-  ambulance_selector:
-    class: RandomSelector
-    classpath: ems.algorithms.selection.dispatch_random
+  ambulances: "@ambulances"
+  cases: "@cases"
+  ambulance_selector: "@ambulance_selector"
